+++ pkg/tests/apis/provisioning/common/testing.go
@@ -1520,7 +1520,7 @@ func WithProvisioningPublicRootURL(url string) GrafanaOption {
 // WithProvisioningWebhookRateLimitRPS sets [provisioning] webhook_rate_limit_rps,
 // enabling the per-client webhook rate limiter at the given sustained rate. The
 // limiter's burst is twice the rate. A value <= 0 leaves the limiter disabled.
-// Without WithProvisioningWebhookTrustedIPHeader the limiter keys on the TCP peer.
+// With no trusted IP header configured the limiter keys on the TCP peer.
 func WithProvisioningWebhookRateLimitRPS(rps int) GrafanaOption {
 	return func(opts *testinfra.GrafanaOpts) {
 		opts.ProvisioningWebhookRateLimitRPS = rps
@@ -1528,8 +1528,8 @@ func WithProvisioningWebhookRateLimitRPS(rps int) GrafanaOption {
 }
 
 // WithProvisioningWebhookTrustedIPHeader sets [provisioning]
-// webhook_trusted_ip_header, naming the header the rate limiter keys on. When
-// unset, the limiter keys on the real TCP peer instead.
+// webhook_trusted_ip_header, naming the header whose value the rate limiter keys
+// on. When unset, the limiter keys on the real TCP peer instead.
 func WithProvisioningWebhookTrustedIPHeader(header string) GrafanaOption {
 	return func(opts *testinfra.GrafanaOpts) {
 		opts.ProvisioningWebhookTrustedIPHeader = header
@@ -3176,13 +3176,16 @@ func RunGrafanaWithGitServer(t *testing.T, options ...GrafanaOption) *GitTestHel
 func runGrafanaWithGitServerShared(t *testing.T, options ...GrafanaOption) (*GitTestHelper, func()) {
 	t.Helper()
 
-	// Bind the container lifecycle to a standalone context, not t.Context(): the
-	// shared server must outlive the single test that happens to trigger init via
-	// sync.Once. If it inherited that test's context, the context would cancel when
-	// that test finished, dropping the testcontainers reaper connection and letting
-	// Ryuk tear down the Gitea container while later tests still need it (surfacing
-	// as "No such container" on CreateUser). Teardown is handled by shutdown() below.
-	gitServer, err := gittest.NewServer(context.Background(), gittest.WithLogger(gittest.NewWriterLogger(os.Stderr)))
+	// Bind the container lifecycle to a non-cancellable context: the shared server
+	// must outlive the single test that happens to trigger init via sync.Once. If
+	// it inherited that test's cancellation, the context would cancel when that
+	// test finished, dropping the testcontainers reaper connection and letting Ryuk
+	// tear down the Gitea container while later tests still need it (surfacing as
+	// "No such container" on CreateUser). WithoutCancel keeps t.Context()'s values
+	// (deadline aside) while stripping cancellation. Teardown is handled by
+	// shutdown() below.
+	ctx := context.WithoutCancel(t.Context())
+	gitServer, err := gittest.NewServer(ctx, gittest.WithLogger(gittest.NewWriterLogger(os.Stderr)))
 	require.NoError(t, err, "failed to start git server")
 
 	allOpts := append([]GrafanaOption{WithRepositoryTypes([]string{"git"})}, options...)
